backend/requirements.txt
fastapi==0.85.0
uvicorn==0.18.2
pymongo==4.2.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==0.20.0
email-validator==1.2.1
